Prevent the derivative kick
When there is a sudden change in the set point, there is a spike in the
output because the change of the set point causes an instantaneous
change in the error. By assuming a constant set point, we can change
the controller to give a similar output, even when the set point does
change, without the output spike.